create compliant

+++ feedback/src/main/java/com/sprng/feedback/service/FeedBackService.java
@@ -1,20 +1,30 @@
 package com.sprng.feedback.service;
 
+import com.sprng.feedback.model.Complaint;
 import com.sprng.feedback.model.FeedBack;
+import com.sprng.feedback.repository.CompliantRepository;
 import com.sprng.feedback.repository.FeedBackRepository;
 import com.sprng.library.exception.IshopResponseException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+
+import java.time.ZonedDateTime;
 
 @Service
 public class FeedBackService {
     private FeedBackRepository feedBackRepository;
+    private ClientCustomerAndTraderService clientCustomerAndTraderService;
+    private CompliantRepository compliantRepository;
 
-    @Autowired
-    public FeedBackService(FeedBackRepository feedBackRepository) {
+    public FeedBackService(FeedBackRepository feedBackRepository,
+                           ClientCustomerAndTraderService clientCustomerAndTraderService,
+                           CompliantRepository compliantRepository) {
         this.feedBackRepository = feedBackRepository;
+        this.clientCustomerAndTraderService = clientCustomerAndTraderService;
+        this.compliantRepository = compliantRepository;
     }
 
     public Mono<FeedBack> createFeedBack(FeedBack feedBack) {
@@ -36,6 +46,8 @@ public class FeedBackService {
     public Mono<FeedBack> deleteFeedBackById(int feedBackID) {
         Mono<FeedBack> feedBack = feedBackRepository.findById(feedBackID)
                 .flatMap((feedback) -> {
+                    if (feedback == null)
+                        return Mono.error(new IshopResponseException("Отзыв по указанному id не найден "));
                     feedback.delete();
                     return feedBackRepository.save(feedback);
                 });
@@ -43,6 +55,38 @@ public class FeedBackService {
     }
 
 
+    public Mono<Complaint> leaveComplaint(Complaint complaint) {
+        clientCustomerAndTraderService
+                .existsByCustomerAndProductTrader(complaint.getCustomerId(),
+                        complaint.getProductTraderId()).flatMap((booleanValue) -> {
+                    if (booleanValue) {
+                        complaint.setCreatedAt(ZonedDateTime.now());
+                       return compliantRepository.save(complaint);
+                    } else {
+                     return Mono.error(new IshopResponseException("Customer or Trader does not exist"));
+                    }
+                });
+        return null;
+
+    }
+
+    public Flux<Complaint> getUnprocessedComplaint() {
+
+       return compliantRepository.findAllByStatusComplaintFalseOrderByCreatedAtDesc();
+    }
+
+    public Mono<String> completeComplaintProcessing(int compliantId) {
+       return compliantRepository.findById(compliantId)
+               .flatMap((compliant)-> {
+                   compliant.setStatusComplaint(true);
+                   compliantRepository.save(compliant);
+                   return Mono.just("Success");
+               });
+    }
+
+    public Mono<Integer> complaintCount(int productTraderId) {
+      return compliantRepository.countByProductTraderId(productTraderId);
+    }
 }
